refactor(layout): 레이아웃 관련 코드 디렉터리 변경
- Header, Layout 등
